requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyjwt
cryptography
sqlalchemy
prometheus-client
psycopg2
prometheus-fastapi-instrumentator
uvicorn[standard]
pytest
httpx
pytest-cov
ruff